Move helper modules into separate dir, create callaws file to create server or browser client, create pinpoint sender module
