+++ packages/analytics-plugin-aws-pinpoint/src/pinpoint/helpers/create-event-queue.js
