+++ packages/analytics-plugin-aws-pinpoint/src/pinpoint/helpers/events.js
